Use CanCreateSymbolicLink instead of IsSymLinkSupported
