+++ src/libraries/Common/tests/System/IO/ReparsePointUtilities.cs
@@ -10,14 +10,10 @@ This is meant to contain useful utilities for IO related work in ReparsePoints
 #define DEBUG
 
 using System;
-using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
-using System.Linq;
 using System.Runtime.InteropServices;
-using System.Text;
 using System.Threading.Tasks;
-using Xunit;
 
 public static partial class MountHelper
 {
@@ -35,77 +31,7 @@ public static partial class MountHelper
     // Helper for ConditionalClass attributes
     internal static bool IsSubstAvailable => PlatformDetection.IsSubstAvailable;
 
-    /// <summary>
-    /// In some cases (such as when running without elevated privileges),
-    /// the symbolic link may fail to create. Only run this test if it creates
-    /// links successfully.
-    /// </summary>
-    internal static bool CanCreateSymbolicLinks => s_canCreateSymbolicLinks.Value;
 
-    private static readonly Lazy<bool> s_canCreateSymbolicLinks = new Lazy<bool>(() =>
-    {
-        bool success = true;
-
-        // Verify file symlink creation
-        string path = Path.GetTempFileName();
-        string linkPath = path + ".link";
-        success = CreateSymbolicLink(linkPath: linkPath, targetPath: path, isDirectory: false);
-        try { File.Delete(path); } catch { }
-        try { File.Delete(linkPath); } catch { }
-
-        // Verify directory symlink creation
-        path = Path.GetTempFileName();
-        linkPath = path + ".link";
-        success = success && CreateSymbolicLink(linkPath: linkPath, targetPath: path, isDirectory: true);
-        try { Directory.Delete(path); } catch { }
-        try { Directory.Delete(linkPath); } catch { }
-
-        // Reduce the risk we accidentally stop running these altogether
-        // on Windows, due to a bug in CreateSymbolicLink
-        if (!success && PlatformDetection.IsWindows)
-            Assert.True(!PlatformDetection.IsWindowsAndElevated);
-
-        return success;
-    });
-
-    /// <summary>Creates a symbolic link using command line tools.</summary>
-    public static bool CreateSymbolicLink(string linkPath, string targetPath, bool isDirectory)
-    {
-        // It's easy to get the parameters backwards.
-        Assert.EndsWith(".link", linkPath);
-        if (linkPath != targetPath) // testing loop
-            Assert.False(targetPath.EndsWith(".link"), $"{targetPath} should not end with .link");
-
-#if NETFRAMEWORK
-        bool isWindows = true;
-#else
-        if (OperatingSystem.IsIOS() || OperatingSystem.IsTvOS() || OperatingSystem.IsMacCatalyst() || OperatingSystem.IsBrowser()) // OSes that don't support Process.Start()
-        {
-            return false;
-        }
-
-        bool isWindows = OperatingSystem.IsWindows();
-#endif
-        Process symLinkProcess = new Process();
-        if (isWindows)
-        {
-            symLinkProcess.StartInfo.FileName = "cmd";
-            symLinkProcess.StartInfo.Arguments = string.Format("/c mklink{0} \"{1}\" \"{2}\"", isDirectory ? " /D" : "", linkPath, targetPath);
-        }
-        else
-        {
-            symLinkProcess.StartInfo.FileName = "/bin/ln";
-            symLinkProcess.StartInfo.Arguments = string.Format("-s \"{0}\" \"{1}\"", targetPath, linkPath);
-        }
-        symLinkProcess.StartInfo.UseShellExecute = false;
-        symLinkProcess.StartInfo.RedirectStandardOutput = true;
-
-        symLinkProcess.Start();
-
-        symLinkProcess.WaitForExit();
-
-        return (symLinkProcess.ExitCode == 0);
-    }
 
     /// <summary>On Windows, creates a junction using command line tools.</summary>
     public static bool CreateJunction(string junctionPath, string targetPath)
