+++ src/libraries/System.IO.FileSystem.Watcher/tests/FileSystemWatcher.SymbolicLink.cs
@@ -7,13 +7,13 @@ using Xunit.Sdk;
 
 namespace System.IO.Tests
 {
-    [ConditionalClass(typeof(MountHelper), nameof(MountHelper.CanCreateSymbolicLinks))]
+    [ConditionalClass(typeof(PlatformDetection), nameof(PlatformDetection.CanCreateSymbolicLinks))]
     public class SymbolicLink_Changed_Tests : FileSystemWatcherTest
     {
         private string CreateSymbolicLinkToTarget(string targetPath, bool isDirectory, string linkPath = null)
         {
             linkPath ??= GetRandomLinkPath();
-            Assert.True(MountHelper.CreateSymbolicLink(linkPath, targetPath, isDirectory));
+            Assert.True(SymbolicLinkHelper.CreateSymbolicLink(linkPath, targetPath, isDirectory));
 
             return linkPath;
         }
