+++ src/libraries/System.IO.FileSystem/tests/Directory/GetFiles.cs
@@ -16,7 +16,7 @@ namespace System.IO.Tests
             return Directory.GetFiles(path);
         }
 
-        [ConditionalFact(typeof(MountHelper), nameof(MountHelper.CanCreateSymbolicLinks))]
+        [ConditionalFact(typeof(PlatformDetection), nameof(PlatformDetection.CanCreateSymbolicLinks))]
         public void EnumerateWithSymLinkToFile()
         {
             DirectoryInfo containingFolder = Directory.CreateDirectory(GetTestFilePath());
@@ -26,7 +26,7 @@ namespace System.IO.Tests
             targetFile.Create().Dispose();
 
             string linkPath = Path.Combine(containingFolder.FullName, GetRandomLinkName());
-            Assert.True(MountHelper.CreateSymbolicLink(linkPath, targetFile.FullName, isDirectory: false));
+            Assert.True(SymbolicLinkHelper.CreateSymbolicLink(linkPath, targetFile.FullName, isDirectory: false));
 
             Assert.True(File.Exists(linkPath));
             Assert.Equal(1, GetEntries(containingFolder.FullName).Count());
