+++ src/libraries/System.IO.FileSystem/tests/FileInfo/SymbolicLinks.cs
@@ -6,7 +6,7 @@ using Xunit;
 
 namespace System.IO.Tests
 {
-    [ConditionalClass(typeof(MountHelper), nameof(MountHelper.CanCreateSymbolicLinks))]
+    [ConditionalClass(typeof(PlatformDetection), nameof(PlatformDetection.CanCreateSymbolicLinks))]
     public class FileInfo_SymbolicLinks : BaseSymbolicLinks_FileSystemInfo
     {
         protected override bool IsDirectoryTest => false;
