+++ src/libraries/System.IO.FileSystem/tests/Junctions.Windows.cs
@@ -6,7 +6,7 @@ using Xunit;
 namespace System.IO.Tests
 {
     [PlatformSpecific(TestPlatforms.Windows)]
-    [ConditionalClass(typeof(MountHelper), nameof(MountHelper.CanCreateSymbolicLinks))]
+    [ConditionalClass(typeof(PlatformDetection), nameof(PlatformDetection.CanCreateSymbolicLinks))]
     public class Junctions : BaseSymbolicLinks
     {
         private DirectoryInfo CreateJunction(string junctionPath, string targetPath)
